Add jdk.crypto.cryptoki and jdk.random
If jlink:ing an application which depends on chariot,
that application image will now automatically include jdk.crypto.cryptoki
(and jdk.crypto.ec) and jdk.random.
Those are needed when connecting to Lichess over HTTPS (SSL handshake),
and when generating code verifier for OAuth PKCE support.
(It is possible to build an image without them, but the application
would fail in runtime when trying to connect or trying to use OAuth
PKCE)