Update build with more info
Currently see a different hash on the built jars on GitHub Actions vs
building the jars locally - so trying to find out what is missing from
the "Reproducible Build" concept.

One known difference is that locally I have,
strip-nondeterminism version 1.12
And on GitHub Actions runner the version is
strip-nondeterminism version 1.7

But trying 1.7 locally in a docker container did produce same hash that
my local 1.12 did - so might be something else I'm unaware of (or I
just messed up in testing).