Add convenience method for OAuth2 PKCE client
First time the user will be asked to go to Lichess site to grant the
application access for the requested scope/s.
Consecutive times, the granted token will be reused and
no interaction will be necessary.

Usage:

    var client = Client.reuseOrInitialize(params -> params
        .prefs("my-preferences")
        .scopes(Client.Scope.tournament_write));

Which is internally implemented with:

    var prefs = Preferences.userRoot().node(preferences);
    var client = Client.load(prefs);
    if (client instanceof ClientAuth auth &&
        auth.account().scopes().containsAll(Set.of(scopes)))
        return auth;

    var oauth = client.account().oauthPKCE(scopes);
    System.out.println("\n\nGrant access at URL:\n" + oauth.url() + "\n");
    var clientAuth = Client.load(prefs, auth -> auth.auth(oauth.token().get()));
    clientAuth.store(prefs);

    return clientAuth;

Alternative,
shorter but needs preparation and handling of a personal token:

    var client = Client.auth("my-token");