Include auth token in reqeuests if present
Some endpoints don't need authorization tokens, but can offer speedups
in throttling when one is used.
Another reason is for the Studies API, which has methods to export
studies. It works fine for public studies, but if a study is made
private, then a token with scope study:read must be used. And even if
the client was created using such a token, it still wasn't sent because
the endpoint doesn't strictly require them.
This commit always includes the token if the client has been provided
with one.