Revert "Filter out empty strings"
This reverts commit d2dd243.

Sometimes empty strings are valid input, so can't filter them out.
The problem the reverted commit tried to stop was when trying to map
such an empty string into a ChallengeResult. When that failed, it
became an Unmapped entry. And that failed miserably with
ClassCastException, as that class wasn't handled.
So reverting also that concept for now - and using a null result for
now, as nulls are currently being handled (discarded).
Future fix to consider could maybe be to have the mapping function
be typed <String, Optional<T>> and use Optional.empty() instead of
null, or something.