Run app container with hostname app
If not giving a hostname, the hostname will be random, something like
fe12227b6e17 - it's nicer with "app".
If writing a integration test for broadcasts with sync url,
it's nice if lila can be told that the sync url is for the host "app"
instead of the random "fe12227b6e17".