Set header "accept-language: en"
Currently there exists code which parses tournament conditons/verdicts:

chariot.internal.modeladapter.TournamentAdapter
    Condition parseEntryCondition(String condition)

Recently (or always but unnoticed) the tournament conditions are
translated to the language of the user account querying the API. The
code which performs the parsing currently expects english. In order to
make the code happy, chariot will now request english with the
'accept-language' header.